Concert page models:
This commit fills in functionality with Concert/Performance relations.
It extends the Wagtail admin form on Performances and People to hide the
Title field from the user and generate a slug. It also adds
`verbose_name` and `__str__` to models to get sensible names in the
admin.

It also commits a number of migrations for all the changes in
`models.py` - these will be consolidated in a future commit.

Lastly, it adds settings for local Postgres so there won't be any
surprises when this goes live.