Fixes save bug in image import, implements an image fix mgmt cmd
This fix-up command is needed to update images with crops from the
drupal site. The original import script had a bug that didn't save the
images crops after importing. It also had a bug such where, if the
particular crop wasn't present, it would return none. If this happens,
this script will fall back to the first image link it can find, and then
apply feature detection to determine a crop.